requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
twilio
requests
sqlalchemy
aiosqlite
httpx
pydantic[email]
pandas
websockets
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
